requirements.txt
Pillow>=11.0.0
rembg>=2.0.0
tkinter-tooltip==2.1.0

streamlit
onnxruntime

